requirements.txt
dropbox
easyocr
git+https://github.com/openai/CLIP.git
open_clip_torch
faiss-cpu